refactor: clean up dead code in apply_cmd.py after Phase 1 extraction
- Remove duplicate functions that were moved to execution/apply.py:

  - _validate_manifest_content()

  - _discover_and_validate_files()

  - _execute_planned_commands()

- Clean up unused imports: glob, Path, yaml, run_kubectl_with_yaml, CommandAction

- Update function call to use execution module: run_intelligent_apply_workflow()

- Achieve 100% test coverage for apply_cmd.py (39 statements, 0 missed)

- All tests passing with simplified, focused command handler

This completes Phase 1 cleanup - apply_cmd.py now has clean separation

of concerns matching the edit_cmd.py pattern.
